Optionally run customize script in chroot and headless.
If the new "--customize-in-chroot" option is specified, run the
customize script in chroot.  Whether or not that new option is
specified, the customize script is now copied to the root of the
created image to make builds reproducible.

If the customize script fails when trying to open a TTY device, assume
no TTY devices are available and complete the build in headless mode.
This headless-mode fallback is required for automated builders, like
buildbot, in which no TTY device is available during the build process.